BUILD: moves CL and TL to common "components" dir
    according to WG discussions, since TL and CL inherit from common
    "base" interface the code is moved to the common dir called "components"
