+++ src/Makefile.am
@@ -1,7 +1,7 @@
 #
 # Copyright (C) Mellanox Technologies Ltd. 2020.  ALL RIGHTS RESERVED.
 #
-SUBDIRS = cl/basic
+SUBDIRS = components/cl/basic
 
 lib_LTLIBRARIES  = libucc.la
 noinst_LIBRARIES =
@@ -16,28 +16,30 @@ nobase_dist_libucc_la_HEADERS = 	\
 	ucc/api/ucc_version.h           \
 	ucc/api/ucc_status.h
 
-noinst_HEADERS =             \
-	core/ucc_global_opts.h   \
-	core/ucc_lib.h           \
-	core/ucc_context.h       \
-	utils/ucc_compiler_def.h \
-	utils/ucc_log.h          \
-	utils/ucc_parser.h       \
-	utils/ucc_component.h    \
-	utils/ucc_datastruct.h   \
-	utils/ucc_math.h         \
-	cl/ucc_cl.h              \
-	cl/ucc_cl_log.h          \
-	cl/ucc_base_iface.h
+noinst_HEADERS =                     \
+	core/ucc_global_opts.h           \
+	core/ucc_lib.h                   \
+	core/ucc_context.h               \
+	utils/ucc_compiler_def.h         \
+	utils/ucc_log.h                  \
+	utils/ucc_parser.h               \
+	utils/ucc_component.h            \
+	utils/ucc_datastruct.h           \
+	utils/ucc_math.h                 \
+	components/base/ucc_base_iface.h \
+	components/cl/ucc_cl.h           \
+	components/cl/ucc_cl_log.h
 
-libucc_la_SOURCES =        \
-	core/ucc_lib.c         \
-	core/ucc_constructor.c \
-	core/ucc_global_opts.c \
-	core/ucc_version.c     \
-	core/ucc_context.c     \
-	utils/ucc_component.c  \
-	cl/ucc_cl.c            \
-	cl/ucc_base_iface.c
+
+
+libucc_la_SOURCES =                  \
+	core/ucc_lib.c                   \
+	core/ucc_constructor.c           \
+	core/ucc_global_opts.c           \
+	core/ucc_version.c               \
+	core/ucc_context.c               \
+	utils/ucc_component.c            \
+	components/base/ucc_base_iface.c \
+	components/cl/ucc_cl.c
 
 libucc_ladir = $(includedir)
