+++ src/components/base/ucc_base_iface.c
