+++ src/components/cl/basic/cl_basic_lib.c
@@ -5,7 +5,6 @@
  */
 
 #include "cl_basic.h"
-#include "cl/ucc_cl_log.h"
 #include "utils/ucc_malloc.h"
 
 UCC_CLASS_INIT_FUNC(ucc_cl_basic_lib_t, const ucc_base_lib_params_t *params,
