+++ src/core/ucc_lib.c
@@ -10,7 +10,7 @@
 #include "utils/ucc_log.h"
 #include "utils/ucc_malloc.h"
 #include "utils/ucc_math.h"
-#include "cl/ucc_cl.h"
+#include "components/cl/ucc_cl.h"
 
 UCS_CONFIG_DEFINE_ARRAY(cl_types, sizeof(ucc_cl_type_t),
                         UCS_CONFIG_TYPE_ENUM(ucc_cl_names));
